requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
clean-fid==0.1.35
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
hydra-core==1.3.2
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
lightning==2.6.0
lightning-utilities==0.15.2
Markdown==3.10
MarkupSafe==2.1.5
matplotlib==3.10.8
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
numpy==2.3.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
packaging==25.0
pillow==12.0.0
platformdirs==4.5.1
propcache==0.4.1
protobuf==6.33.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.0
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.3
sentry-sdk==2.48.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8.1
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==2.5.1+cu124
torchmetrics==1.8.2
torchvision==0.20.1+cu124
tqdm==4.67.1
triton==3.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
wandb==0.23.1
Werkzeug==3.1.4
wheel==0.45.1
yarl==1.22.0
